+++ Delta.CertXplorer/src/Delta.CertXplorer.Extensibility/BaseDataHandlerPlugin.cs
@@ -1,0 +1,11 @@
+﻿namespace Delta.CertXplorer.Extensibility
+{
+    public abstract class BaseDataHandlerPlugin : BasePlugin, IDataHandlerPlugin
+    {
+        #region IDataHandlerPlugin Members
+
+        public abstract IDataHandler CreateHandler();
+
+        #endregion
+    }
+}